+++ .gitignore
@@ -3,7 +3,7 @@
 # dependencies
 /.pnp
 .pnp.js
-/node_modules
+node_modules/
 
 # testing
 /coverage
