+++ client/src/App.js
@@ -1,6 +1,6 @@
 import NewUser from "./pages/NewUser";
 import Login from "./pages/Login";
-import NoMatch from "./pages/NoMatch";
+// import NoMatch from "./pages/NoMatch";
 import Home from "./pages/Home";
 import PersonalPage from "./pages/PersonalPage";
 import NewJoke from "./pages/NewJoke";
@@ -51,9 +51,9 @@ const AppRoutes = () => {
           <Route exact path="/newJoke">
              <NewJoke />
            </Route>
-           <Route>
+           {/* <Route>
              <NoMatch />
-           </Route>
+           </Route> */}
           </AuthenticatedRoute>
           <UnauthenticatedRoutes />
         </Switch>
