+++ client/src/pages/PersonalPage.js
@@ -1,5 +1,5 @@
-// import React from "react";
-import React, { useState, useEffect } from "react";
+import React from "react";
+// import React, { useState, useEffect } from "react";
 import CreateNewJokeBtn from "../components/CreateNewJokeBtn";
 import HomePageBtn from "../components/HomePageBtn";
 import LogOutBtn from "../components/LogOutBtn";
@@ -10,32 +10,32 @@ import Footer from "../components/Footer";
 import Container from "../components/Container";
 import Nav from "../components/Nav";
 
-import API from "../utils/API";
+// import API from "../utils/API";
 
 function PersonalPage() {
 
-  // Setting our component's initial state
-  const [jokes, setJokes] = useState([])
+   // Setting our component's initial state
+  // const [jokes, setJokes] = useState([])
   
-  // Load all jokes and store them with setJokes
-  useEffect(() => {
-    console.log("hi")
-    loadJokes()
-  }, [])
+  // // Load user and store them with setJokes
+  // useEffect(() => {
+  //   console.log("hi")
+  //   loadJokes()
+  // }, [])
 
-  // Loads all jokes and sets them to jokes
-  function loadJokes() {
-    const user = localStorage.getItem("userInfo")
-    const userInfo = JSON.parse(user)
-    const userID = userInfo[0]._id
-    console.log("also hi")
-    API.getUsersById(userID)
-      .then(res => 
-        setJokes(res.data)
-      )
-      .catch(err => console.log(err));
-  };
-  
+  // // Loads current user and sets them to jokes
+  // function loadJokes() {
+  //   const user = localStorage.getItem("userInfo")
+  //   const userInfo = JSON.parse(user)
+  //   const userID = userInfo[0]._id
+  //   console.log("also hi")
+  //   API.getUsersById(userID)
+  //     .then(res => 
+  //       setJokes(res.data),
+  //     )
+  //     .catch(err => console.log(err));
+  // };
+
   return <div>
   <Header />
   <Nav>
@@ -45,7 +45,9 @@ function PersonalPage() {
   <OwnJokesBtn />
   </Nav> 
   <Container>
-  <SavedJokeCard jokes = {jokes} />
+  <SavedJokeCard 
+  // jokes = {jokes} 
+  />
   </Container>
   <Footer />
   </div>
